[Sync fork] Syncing changes from oracle/graal/master

+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/truffle/test/strings/TStringOpsCopyConstantStrideTest.java
@@ -28,15 +28,15 @@ import static org.junit.runners.Parameterized.Parameters;
 
 import java.util.List;
 
-import jdk.graal.compiler.nodes.StructuredGraph;
-import jdk.graal.compiler.nodes.graphbuilderconf.GraphBuilderConfiguration;
-import jdk.graal.compiler.options.OptionValues;
-import jdk.graal.compiler.replacements.nodes.ArrayCopyWithConversionsNode;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
+import jdk.graal.compiler.nodes.StructuredGraph;
+import jdk.graal.compiler.nodes.graphbuilderconf.GraphBuilderConfiguration;
+import jdk.graal.compiler.options.OptionValues;
+import jdk.graal.compiler.replacements.nodes.ArrayCopyWithConversionsNode;
 import jdk.vm.ci.code.InstalledCode;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 
@@ -73,7 +73,7 @@ public class TStringOpsCopyConstantStrideTest extends TStringOpsCopyTest {
     @Override
     @Test
     public void testCopy() {
-        ArgSupplier arrayB = () -> new byte[128 + offsetB + (lengthCPY << strideB) + 128];
+        ArgSupplier arrayB = () -> new byte[(int) (128 + offsetB + (lengthCPY << strideB) + 128)];
         constantArgs[3] = strideA;
         constantArgs[7] = strideB;
         testWithNativeExcept(getArrayCopyWithStride(), null, 1 << 5, DUMMY_LOCATION, arrayA, offsetA, strideA, 0, arrayB, offsetB, strideB, 0, lengthCPY);
